Optimize nested comparisons (ticket #2132).
This patch implements a new optimization simplCore/Comparisons, which
keeps track what comparisons have been done and uses that to remove
ones that are guaranteed to be always true or false. For example:
  case x ># y of
    True -> .. x ==# y ..
    ..
the second comparison will always be false. It also tracks possible
values of variables using intervals, so in:
  case x ># 5 of
    True -> .. x ==# 0 ..
    ..
it will detect that the second comparison is always false. Finally it
is also able to use transitivity so here:
  case x ># y of
    True -> .. case y ># z of
      True -> .. x ># z ..
    ..
it will optimize away the last comparison.