Interpreter.c: use macros to access/modify Sp
This is another step in fixing #13825 and based on D38 by Simon
Marlow.

This commit adds a few macros for accessing and modifying `Sp`
(interpreter stack). This will be useful later on, to allow sub-word
indexing/pushing. (but that will be the next step, this commit should
introduce no changes in behavior)

Signed-off-by: Michal Terepeta <michal.terepeta@gmail.com>